fixes: Bug 1 — scripts/run_experiment.py (CRITICAL, now fixed) When running a dc sweep (--dc 8 16 24 32 48) with cov_pca or cov_unsupervised configs, the pretrained checkpoint path was wrong for any dc ≠ 32. E.g. dc=16 → looked for pca_pooler_dc32_dc16.pt instead of pca_pooler_dc16.pt. Fix: strip existing _dcNN suffix with a regex before appending the new one.
Bug 2 — src/sop/pooling/covariance.py + covariance_pca.py (AMP performance)
Under AMP on T4, bmm() outputs fp16 but lengths was float32. Dividing fp16÷fp32 silently upcasts the result to float32, throwing away the fp16 speedup. Fix: .to(XL.dtype) / .to(Xp.dtype) on lengths before the division.

Bug 3 — src/sop/probes/train_loop.py (deprecation warning)
torch.cuda.amp.GradScaler is deprecated in PyTorch ≥2.4 (which Colab currently ships). Fixed to torch.amp.GradScaler(cuda, ...).

Bug 4 — src/sop/analysis/plots.py (visualization)
ax.fill_between(ax.get_xlim(), ...) called before any data was plotted, so the x-limits were still the matplotlib default (0, 1). The mean±std band would be drawn off-screen. Fixed with axhspan which spans the full x-range regardless of when it's called.

The core math (Frobenius identity, PCA weighting, covariance bilinear form, masking) is all correct — and the test suite is solid.

+++ src/sop/pooling/covariance.py
@@ -67,7 +67,7 @@ class CovariancePooler(Pooler):
         XL = self.proj_l(Xm)                                      # [B, L, dc]
         XR = self.proj_r(Xm)                                      # [B, L, dc]
 
-        C = torch.bmm(XL.transpose(1, 2), XR) / lengths.squeeze(-1).unsqueeze(-1)
+        C = torch.bmm(XL.transpose(1, 2), XR) / lengths.squeeze(-1).unsqueeze(-1).to(XL.dtype)
         out = C.flatten(1)                                        # [B, dc²]
         return out.squeeze(0) if single else out
